print true if given numbers r coprimes else false
We are given two integers a and b. For those two integers we look for positive integers which divide both a and b (without leaving a remainder). If there does not exist such an integer besides 1, we call a and b coprime.
Input
Line 1: A positive integer a.
Line 2: A positive integer b.
Output
The string "true" when the given integers are coprime or "false" when they are not.
Constraints
1 ≤ a, b ≤ 2×10^7
Example
Input
5
6
Output
true