rainbow is accepted or defined
All the allowed colors should be present in a rainbow request, in order.
Allowed colors in the wrong order and not allowed colors are considered illegal characters.
Only characters between A and Z are considered colors, everything else is ignored.
Input
Line 1: A string with the allowed characters
Line 2: A string to check
Output
If the requested rainbow is legal: Rainbow accepted! 0
else: Rainbow denied! n
where n is the number of illegal characters in the request
Constraints
0 < Length of rainbow < 100
Example
Input
RAPFJ
R99pAaPF!J
Output
Rainbow accepted! 0