print how many groups with no.of girls & boys
You have x girls and y boys. You want to create groups with boys and girls. Each group must have the same number of boys and the same number girls. What is the maximum number of groups that you can make? How many boys and how many girls there would be in each group?

For exemple if you have 3 girls and 6 boys, you can make a maximum of 3 groups with 1 girl and 2 boys in each group.
Input
Line 1: An integer x the numbers of girls and an integer y the numbers of boys separated by a space.
Output
Line 1: The maximum number of groups you can make.
Line 2: The number of girls and the number boys in each group separated by a space.
Constraints
10 ≤ x y ≤ 100
Example
Input
18 30
Output
6
3 5