print the next Fibonacci number
Adam is trying out various sequences similar to Fibonacci sequences. He starts with two numbers, and keeps iteratively getting the next number in the sequence by adding the two previous numbers.

After a point, Adam always gets stuck, and he looks to you for help in finding the next number in the sequence. Given the sequence of numbers generated by Adam, find the next Fibonacci-like number in the sequence.

Unfortunately, Adam sometimes makes a mistake while creating this sequence. In such situations, you need to tell Adam that his sequence is invalid and he needs to fix it before continuing.

Note: The first two numbers in the sequence define the sequence, and will never be considered wrong.

If the sequence of numbers is a valid Fibonacci-like sequence, print the next number in the sequence. If the sequence is invalid, print the string "Invalid".
Input
Line 1: An integer N for the number of values that follow
Line 2: N space-separated integers
Output
One integer, the next number in the sequence OR the string "Invalid"
Constraints
3 ≤ N ≤ 15
-1000 ≤ each number in sequence ≤ 1000

Output must either be "Invalid" or an integer between -1000 and 1000
Example
Input
6
5 8 13 21 34 55
Output
89